Network stuff

+++ Wildfire ICS Assist/OrganizationalChartForm.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.VisualBasic;
+﻿
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -10,6 +10,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using WF_ICS_ClassLibrary;
 using WF_ICS_ClassLibrary.EventHandling;
 using WF_ICS_ClassLibrary.Models;
 using WF_ICS_ClassLibrary.Utilities;
@@ -38,8 +39,14 @@ namespace Wildfire_ICS_Assist
 
             Program.wfIncidentService.ICSRoleChanged += Program_ICSRoleChanged;
             Program.wfIncidentService.OrganizationalChartChanged += Program_OrgChartChanged;
+            Program.wfIncidentService.OpPeriodChanged += Program_OpPeriodChanged;
+
 
             chkIncludeContacts.Checked = Program.generalOptionsService.GetOptionsBoolValue("IncludeOrgContactsInIAP");
+        }
+        private void Program_OpPeriodChanged(IncidentOpPeriodChangedEventArgs e)
+        {
+            PopulateTree();
         }
 
         private void Program_OrgChartChanged(OrganizationChartEventArgs e)
@@ -51,7 +58,7 @@ namespace Wildfire_ICS_Assist
         }
         private void Program_ICSRoleChanged(ICSRoleEventArgs e)
         {
-            if (e.item != null && CurrentOrgChart.AllRoles.Any(o => o.RoleID == e.item.RoleID)) { PopulateTree(e.item); }
+            if (e.item != null && CurrentOrgChart.ActiveRoles.Any(o => o.RoleID == e.item.RoleID)) { PopulateTree(e.item); }
             else { PopulateTree(); }
         }
 
@@ -115,7 +122,7 @@ namespace Wildfire_ICS_Assist
 
         private void AddCurrentChild(Guid parentId, TreeNodeCollection nodes)
         {
-            var rows = CurrentOrgChart.AllRoles.Where(o => o.ReportsTo == parentId).ToList();
+            var rows = CurrentOrgChart.ActiveRoles.Where(o => o.ReportsTo == parentId).ToList();
 
             foreach (var row in rows)
             {
@@ -225,12 +232,29 @@ namespace Wildfire_ICS_Assist
             ICSRole role = (ICSRole)(treeOrgChart.SelectedNode.Tag);
             if (role.AllowDelete)
             {
-                DialogResult dr = MessageBox.Show(Properties.Resources.SureDelete, Properties.Resources.SureDeleteTitle, MessageBoxButtons.YesNo);
+                //check if there are subordinate roles
+                if (Program.CurrentOrgChart.ActiveRoles.Any(o => o.ReportsTo == role.RoleID))
+                {
+                    MessageBox.Show(Properties.Resources.DeleteSubordinateRoles);
+                }
+                else
+                {
+                    DialogResult dr = MessageBox.Show(Properties.Resources.SureDelete, Properties.Resources.SureDeleteTitle, MessageBoxButtons.YesNo);
+                    if (dr == DialogResult.Yes)
+                    {
+                        Program.wfIncidentService.DeleteICSRole(role, Program.CurrentOpPeriod);
+                    }
+                }
+            }
+            else if (role.AllowEditName)
+            {
+                DialogResult dr = MessageBox.Show(Properties.Resources.RenameInsteadOfDeleteRole, Properties.Resources.RenameTitle, MessageBoxButtons.YesNo, MessageBoxIcon.Question);
                 if(dr == DialogResult.Yes)
                 {
-                    Program.wfIncidentService.DeleteICSRole(role, Program.CurrentOpPeriod);
+                    openRoleForEdit(role);
                 }
-            } else { MessageBox.Show(Properties.Resources.ProtectedRole); }
+            }
+            else { MessageBox.Show(Properties.Resources.ProtectedRole); }
 
         }
 
@@ -299,7 +323,33 @@ namespace Wildfire_ICS_Assist
 
         private void rbIncidentCommander_CheckedChanged(object sender, EventArgs e)
         {
-            if (rbIncidentCommander.Checked) { CurrentOrgChart.SwitchToSingleIC(); }
+            if (rbIncidentCommander.Checked)
+            {
+                if (CurrentOrgChart.HasFilledUnifiedCommandRoles)
+                {
+                    DialogResult dr = MessageBox.Show(Properties.Resources.NoSwitchToICWithUCRolesFilled, Properties.Resources.ClearUnifiedCommandRolesTitle, MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                    if (dr == DialogResult.Yes)
+                    {
+                        ICSRole uc2 = CurrentOrgChart.ActiveRoles.FirstOrDefault(o => o.RoleID == Globals.UnifiedCommand2ID);
+                        if (uc2 != null)
+                        {
+                            CurrentOrgChart.UnassignThisAndSubordinateRoles(uc2);
+                        }
+
+                        ICSRole uc3 = CurrentOrgChart.ActiveRoles.FirstOrDefault(o => o.RoleID == Globals.UnifiedCommand3ID);
+                        if (uc3 != null)
+                        {
+                            CurrentOrgChart.UnassignThisAndSubordinateRoles(uc2);
+                        }
+                        CurrentOrgChart.SwitchToSingleIC();
+                    }
+                    else
+                    {
+                        rbUnifiedCommand.Checked = true;
+                    }
+                }
+                else { CurrentOrgChart.SwitchToSingleIC(); }
+            }
         }
 
         private void rbUnifiedCommand_CheckedChanged(object sender, EventArgs e)
@@ -332,7 +382,7 @@ namespace Wildfire_ICS_Assist
 
 
 
-                string csv = OrgChartTools.OrgChartToCSV(CurrentOrgChart.AllRoles, delimiter);
+                string csv = OrgChartTools.OrgChartToCSV(CurrentOrgChart.ActiveRoles, delimiter);
                 try
                 {
                     System.IO.File.WriteAllText(exportPath, csv);
